+++ Src/DSInternals.PowerShell/Commands/Replication/GetADReplBackupKeyCommand.cs
@@ -1,16 +1,13 @@
 ﻿namespace DSInternals.PowerShell.Commands
 {
-    using DSInternals.Common.Data;
-    using DSInternals.PowerShell.Properties;
-    using System;
     using System.Management.Automation;
-    using System.Security.Principal;
+    using DSInternals.Common.Data;
 
     [Cmdlet(VerbsCommon.Get, "ADReplBackupKey")]
     [OutputType(typeof(DPAPIBackupKey))]
     public class GetADReplBackupKeyCommand : ADReplCommandBase
     {
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = false)]
         [ValidateNotNullOrEmpty]
         [Alias("FQDN", "DomainName", "DNSDomainName")]
         public string Domain
@@ -23,17 +20,28 @@
         protected override void BeginProcessing()
         {
             base.BeginProcessing();
-            if(!this.Domain.Contains("."))
+
+            string domainDN;
+            if(string.IsNullOrEmpty(this.Domain))
             {
-                // This is not a hard check, because root domain does not need to have a dot in it.
-                // TODO: Extract as a resource
-                this.WriteWarning("The domain name supplied appears to be a NetBIOS name instead of DNS name.");
+                // Automatically infer DC's domain name.
+                domainDN = this.ReplicationClient.DomainNamingContext;
+            }
+            else
+            {
+                if (!this.Domain.Contains("."))
+                {
+                    // This is not a hard check, because root domain does not need to have a dot in it.
+                    // TODO: Extract as a resource
+                    this.WriteWarning("The domain name supplied appears to be a NetBIOS name instead of DNS name.");
+                }
+
+                // Convert Domain DNS name to distinguished name
+                domainDN = DistinguishedName.GetDNFromDNSName(this.Domain).ToString();
             }
 
             // TODO: Error processing
-            // Convert Domain DNS name to distinguished name
-            var domainDN = DistinguishedName.GetDNFromDNSName(this.Domain);
-            foreach (var backupKey in this.ReplicationClient.GetDPAPIBackupKeys(domainDN.ToString()))
+            foreach (var backupKey in this.ReplicationClient.GetDPAPIBackupKeys(domainDN))
             {
                 this.WriteObject(backupKey);
             }