+++ Src/DSInternals.Replication.Interop/DrsConnection.h
@@ -3,6 +3,8 @@
 #include "drsr_alloc.h"
 #include <string>
 
+#define nameof(parameterName) #parameterName
+
 namespace DSInternals
 {
 	namespace Replication
@@ -25,12 +27,14 @@ namespace DSInternals
 				literal size_t DefaultMaxBytes = 8 * 1024 * 1024; // 8MB
 				literal DWORD DefaultReplEpoch = 0;
 				literal String^ UpnSeparator = "@";
+				literal String^ DummyName = "NULL";
 				// The default session key is "{B84290B0-0C3E-4542-B409-A96DF8DE3D93}"
 				static initonly cli::array<byte>^ DefaultSessionKey = gcnew cli::array<byte> { 0xb8, 0x42, 0x90, 0xb0, 0x0c, 0x3e, 0x45, 0x42, 0xb4, 0x09, 0xa9, 0x6d, 0xf8, 0xde, 0x3d, 0x93 };
 
 				cli::array<byte>^ _sessionKey;
 				Guid _clientDsa;
 				Guid _serverSiteObjectGuid;
+				Guid _configurationObjectGuid;
 				DRS_EXT _serverCapabilities;
 				DWORD _serverReplEpoch;
 				SecurityCallback^ _securityCallback;
@@ -42,6 +46,10 @@ namespace DSInternals
 					cli::array<byte>^ get();
 				}
 				property Guid ServerSiteGuid
+				{
+					Guid get();
+				}
+				property Guid ConfigurationPartitionGuid
 				{
 					Guid get();
 				}
@@ -58,6 +66,14 @@ namespace DSInternals
 				String^ ResolveDistinguishedName(Guid objectGuid);
 				Guid ResolveGuid(NTAccount^ accountName);
 				Guid ResolveGuid(SecurityIdentifier^ objectSid);
+				NTAccount^ ResolveAccountName(String^ distinguishedName);
+				ActiveDirectoryRoleInformation^ ListRoles();
+				array<String^>^ ListSites();
+				array<String^>^ ListNamingContexts();
+				array<String^>^ ListDomains();
+				array<String^>^ ListServersInSite(String^ name);
+				array<String^>^ ListDomainsInSite(String^ name);
+				DomainControllerInformation^ ListInfoForServer(String^ name);
 				bool TestObjectExistence(String^ distinguishedName);
 				bool TestObjectExistence(Guid objectGuid);
 				void WriteNgcKey(String^ distinguishedName, cli::array<byte>^ key);
@@ -72,8 +88,10 @@ namespace DSInternals
 				void Bind(IntPtr rpcHandle);
 				midl_ptr<DRS_MSG_GETCHGREPLY_V9> GetNCChanges(midl_ptr<DRS_MSG_GETCHGREQ_V10>&& request);
 				midl_ptr<DRS_MSG_CRACKREPLY_V1> CrackNames(midl_ptr<DRS_MSG_CRACKREQ_V1>&& request);
-				String^ TryResolveName(String^ name, DS_NAME_FORMAT formatOffered, DS_NAME_FORMAT formatDesired);
-				String^ ResolveName(String^ name, DS_NAME_FORMAT formatOffered, DS_NAME_FORMAT formatDesired);
+				array<String^>^ ResolveNames(array<String^>^ names, DS_NAME_FORMAT formatOffered, DS_NAME_FORMAT formatDesired, bool mustExist);
+				String^ ResolveName(String^ name, DS_NAME_FORMAT formatOffered, DS_NAME_FORMAT formatDesired, bool mustExist);
+				array<String^>^ ListInfo(DS_NAME_FORMAT_EXT infoType);
+				array<String^>^ ListInfo(DS_NAME_FORMAT_EXT infoType, String^ targetName);
 				midl_ptr<DRS_EXTENSIONS_INT> CreateClientInfo();
 				midl_ptr<DRS_MSG_GETCHGREQ_V10> CreateReplicateAllRequest(ReplicationCookie^ cookie, array<ATTRTYP>^ partialAttributeSet, ULONG maxBytes, ULONG maxObjects);
 				midl_ptr<DRS_MSG_GETCHGREQ_V10> CreateReplicateSingleRequest(String^ distinguishedName, array<ATTRTYP>^ partialAttributeSet);